Add class .slide-background-image where relevant
Currently, all slide backgrounds have the class "slide-background".  I'd
like to be able to target those slide backgrounds, which are using an
image (specifically, because I want to reduce the opacity on them).

To do this, we add the class .slide-background-image to all all
instaces of .slide-background where a background image is being used.